refactor app api, move values to separate file, move interface into separate file, change paths
