+++ src/constants/api.ts
@@ -1,0 +1,5 @@
+export const PAGE_NUMBER = 1;
+export const ARTS_FOR_PAGE = 4;
+export const ARTS_FOR_OTHERS = 9;
+export const MAX_RANDOM_PAGE = 100;
+export const API_URL = 'https://api.artic.edu/api/v1';
