+++ src/context/FavoritesContext.tsx
@@ -1,5 +1,5 @@
 import React, { createContext, useContext, useEffect, useState } from 'react';
-import { Artwork } from '@/api/index';
+import { Artwork } from '@/types/Artwork';
 
 interface FavoritesContextType {
   favorites: Artwork[];
