+++ src/types/Artwork.ts
@@ -1,0 +1,13 @@
+export interface Artwork {
+  id: number;
+  title: string;
+  date_display: string;
+  artist_display: string;
+  image_id: string;
+  department_title: string;
+  dimensions: string;
+  credit_line: string;
+  place_of_origin: string;
+  is_public_domain: boolean;
+  artist_title: string;
+}
